backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
greenlet==3.2.3
playwright==1.54.0
psycopg2-binary==2.9.10
pyee==13.0.0
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.14.1
